requirements.txt
Flask==0.10.1
docker-py==0.2.2
redis==2.7.6
requests==1.2.3
humanize==0.5
flask-babel
flask-mail
flask-sqlalchemy
flask-security
sh
